refactored and corrected OpenAPI specs
