+++ pyproject.toml
@@ -140,6 +140,9 @@ lint.per-file-ignores."visionatrix/install_update/*.py" = [
 lint.per-file-ignores."visionatrix/install_update/__init__.py" = [
   "F401",
 ]
+lint.per-file-ignores."visionatrix/routes/*.py" = [
+  "B008",
+]
 lint.extend-per-file-ignores."tests/**/*.py" = [
   "E402",
   "S",
